fix(ws): 修复上游同步后的回调签名
同步上游 v0.1.134 后，HTTP bridge 分支仍使用旧的 AfterTurn 调用签名，导致后端编译失败。

同时修正 usage_log schema 测试中 Size 字段的 int64 断言，匹配 ent migrate column 类型。

验证：git diff --check

+++ backend/ent/schema/usage_log_schema_test.go
@@ -17,7 +17,7 @@ func TestUsageLogRequestIDAllowsWSBillingKeys(t *testing.T) {
 	require.Error(t, runStringValidators(validators, strings.Repeat("x", 256)))
 	require.Error(t, runStringValidators(validators, ""))
 
-	require.Equal(t, 255, requireColumn(t, entmigrate.UsageLogsColumns, "request_id").Size)
+	require.Equal(t, int64(255), requireColumn(t, entmigrate.UsageLogsColumns, "request_id").Size)
 }
 
 func requireStringFieldValidators(t *testing.T, fields []ent.Field, name string) []func(string) error {
